[DeepEP] [fix] Silence -Wc++11-narrowing in moe_distribute_dispatch_v2_ccu.h
perRankDataSize_ was widened uint32_t -> uint64_t, so
`perRankDataSize_ - cntSize/cpSize2` (unsigned long) narrows into the
int64_t DataCopyExtParams initializer field (lines 594/737), which
-Wc++11-narrowing errors on. The failed op-kernel compile left the op
package without binary_info_config.json, breaking aclnnDispatchLayout
at runtime. Wrap with static_cast<int64_t>.

Also apply clang-format to dispatch_layout.h and
grouped_matmul_slice_m_per_token_dequant_swiglu_quant_multistage_workspace.h.

+++ csrc/deepep/ops/op_kernel/moe_distribute_dispatch_v2_ccu.h
@@ -591,7 +591,8 @@ __aicore__ inline void MoeDistributeDispatchA5<TemplateMoeDistributeDispatchA5Ty
     GlobalTensor<uint32_t> sendCntGT;
     sendCntGT.SetGlobalBuffer((__gm__ uint32_t *)(sendBufGM_ + startRank * perRankDataSize_));
     uint32_t cntSize = sizeof(uint32_t);
-    DataCopyExtParams cntParams = {static_cast<uint16_t>(rankPerAiv), cntSize, 0, perRankDataSize_ - cntSize, 0};
+    DataCopyExtParams cntParams = {static_cast<uint16_t>(rankPerAiv), cntSize, 0,
+                                   static_cast<int64_t>(perRankDataSize_ - cntSize), 0};
     DataCopyPad(sendCntGT, cntLT, cntParams);
 }
 
@@ -734,7 +735,7 @@ __aicore__ inline void MoeDistributeDispatchA5<TemplateMoeDistributeDispatchA5Ty
     sendCntGT.SetGlobalBuffer((__gm__ uint32_t *)(sendBufGM_ + (startRank + sharedExpertRankNum_) * perRankDataSize_));
     uint32_t cpSize2 = sizeof(uint32_t) * localExpertNum_;
     DataCopyExtParams cntParams = {static_cast<uint16_t>(rankPerAiv), cpSize2, rankCntSize - cpSize2,
-                                   perRankDataSize_ - cpSize2, 0};
+                                   static_cast<int64_t>(perRankDataSize_ - cpSize2), 0};
     DataCopyPad(sendCntGT, cntLT, cntParams);
 }
